# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bijou AI WhatsApp Enterprise",
    "version": "2.2.0",
    "description": "Enterprise-grade AI customer support for WhatsApp with multi-language support, human escalation, and Google Sheets integration. Built for property agents and SMEs.",
    "contact": {
      "name": "Bijou AI Support Team",
      "email": "support@mybijou.xyz",
      "url": "https://app.mybijou.xyz/static/help.html"
    },
    "license": {
      "name": "Commercial License",
      "url": "https://mybijou.xyz/terms"
    }
  },
  "paths": {
...
    "/api/proactive/schedule": {
      "post": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "Schedule Message",
        "description": "Schedule a message to be sent later.\n\nExample:\n```json\n{\n  \"recipient\": \"60123456789@s.whatsapp.net\",\n  \"message_type\": \"lead_followup\",\n  \"content\": \"Hi! Just following up on your inquiry...\",\n  \"delay_minutes\": 1440\n}\n```",
        "operationId": "schedule_message_api_proactive_schedule_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proactive/campaign": {
      "post": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "Create Campaign",
        "description": "Create a new marketing campaign.\n\nExample:\n```json\n{\n  \"name\": \"New Year Promo\",\n  \"message_template\": \"Happy New Year! Get 20% off...\",\n  \"target_segment\": \"all\",\n  \"scheduled_time\": \"2026-01-01T00:00:00Z\"\n}\n```",
        "operationId": "create_campaign_api_proactive_campaign_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proactive/silence-rule": {
      "post": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "Set Silence Rule",
        "description": "Set up silence detection and auto-reengagement.\n\nExample:\n```json\n{\n  \"silence_days\": 7,\n  \"message_template\": \"We haven't heard from you in a while! How can we help?\"\n}\n```",
        "operationId": "set_silence_rule_api_proactive_silence_rule_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSilenceRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proactive/campaigns": {
      "get": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "List Campaigns",
        "description": "List all campaigns for the authenticated tenant",
        "operationId": "list_campaigns_api_proactive_campaigns_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proactive/scheduled": {
      "get": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "List Scheduled Messages",
        "description": "List all scheduled messages for the authenticated tenant",
        "operationId": "list_scheduled_messages_api_proactive_scheduled_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proactive/scheduled/{message_id}": {
      "delete": {
        "tags": [
          "proactive-messaging"
        ],
        "summary": "Cancel Scheduled Message",
        "description": "Cancel a scheduled message (only if owned by authenticated tenant)",
        "operationId": "cancel_scheduled_message_api_proactive_scheduled__message_id__delete",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          },
          "target_segment": {
            "type": "string",
            "title": "Target Segment"
          },
          "scheduled_time": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recipient_count": {
            "type": "integer",
            "title": "Recipient Count"
          },
          "sent_count": {
            "type": "integer",
            "title": "Sent Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "message_template",
          "target_segment",
          "scheduled_time",
          "status",
          "recipient_count",
          "sent_count",
          "failed_count"
        ],
        "title": "CampaignResponse",
        "description": "Response for campaign"
      },
...
      "CreateCampaignRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          },
          "target_segment": {
            "type": "string",
            "title": "Target Segment"
          },
          "scheduled_time": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Time"
          },
          "recipients": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipients"
          }
        },
        "type": "object",
        "required": [
          "name",
          "message_template",
          "target_segment",
          "scheduled_time"
        ],
        "title": "CreateCampaignRequest",
        "description": "Request to create a campaign"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "scheduled_time": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "recipient",
          "message_type",
          "content",
          "scheduled_time",
          "status"
        ],
        "title": "MessageResponse",
        "description": "Response for scheduled message"
      },
      "ScheduleMessageRequest": {
        "properties": {
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "delay_minutes": {
            "type": "integer",
            "title": "Delay Minutes",
            "default": 0
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "recipient",
          "message_type",
          "content"
        ],
        "title": "ScheduleMessageRequest",
        "description": "Request to schedule a message"
      },
...
      "SetSilenceRuleRequest": {
        "properties": {
          "silence_days": {
            "type": "integer",
            "title": "Silence Days"
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          }
        },
        "type": "object",
        "required": [
          "silence_days",
          "message_template"
        ],
        "title": "SetSilenceRuleRequest",
        "description": "Request to set a silence detection rule"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}